// src/app/api-client/model/permissionStatus.ts
/**
 * Horde Model Reference API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Represents the reviewed permission status for a particular activity.
 */
export const PermissionStatus = {
  Allowed: 'allowed',
  AllowedWithConditions: 'allowed_with_conditions',
  Prohibited: 'prohibited',
  Unknown: 'unknown',
} as const;
export type PermissionStatus = (typeof PermissionStatus)[keyof typeof PermissionStatus];